Use `int' instead of `long' otherwise on 64-bit platforms where `long' is 8-bytes it would not read the byte code properly.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@82674 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/C/bench/meltdump.c
@@ -649,7 +649,7 @@ static  void    read_byte_code (void)
 	{
 		body_id = rlong ();
 
-		if (fwrite (&body_id, sizeof (long), 1, bfp) != 1)
+		if (fwrite (&body_id, sizeof (int), 1, bfp) != 1)
 		{
 			fprintf (stderr,"Write error\n");
 			panic ();
@@ -667,7 +667,7 @@ static  void    read_byte_code (void)
 
 		melt [body_id] = rbuf ((int) bsize);
 
-		if (fwrite (&bsize, sizeof (long), 1, bfp) != 1)
+		if (fwrite (&bsize, sizeof (int), 1, bfp) != 1)
 		{
 			fprintf (stderr,"Write error\n");
 			panic ();
